Add configuration to disable scanning when on battery
This change introduces a new configuration parameter `scan_on_battery`
which makes it possible to skip a scan when running on battery. The
default is `true`, which means that scans will be performed, even when
running on battery. If the parameter is set to `false` the scan will be
skipped if any battery is in state `Discharging`.

Fixes #27